Broke the line-level parser into a whole bunch of smaller, procedural parsers. This is a procedural problem. Thus the classes are really just acting as namespaces.
