+++ src/Core/Gibberish/Parsing/ParseCommentReferences.cs
@@ -1,0 +1,61 @@
+using System;
+using System.Collections.Generic;
+using System.Linq;
+using System.Text.RegularExpressions;
+using Gibberish.AST;
+using JetBrains.Annotations;
+
+namespace Gibberish.Parsing
+{
+	internal static class ParseCommentReferences
+	{
+		[NotNull]
+		public static string ExtractCommentsAndReturnEverythingBeforeThem([NotNull] List<ParseError> errors, [NotNull] string content, [NotNull] List<int> comments)
+		{
+			var statementAndComment = content.Split(
+				new[]
+				{
+					'#'
+				},
+				2);
+			var statement = statementAndComment[0];
+			if (statementAndComment.Length > 1)
+			{
+				statement = statement.TrimEnd();
+				var whitespaceBeforeComment = statementAndComment[0].Substring(statement.Length);
+				if (!whitespaceBeforeComment.Equals("\t\t")) { errors.Add(ParseError.IncorrectCommentSeparator(whitespaceBeforeComment + "#")); }
+				_ExtractCommmentRefs(errors, statementAndComment[1], comments);
+			}
+			return statement;
+		}
+
+		public static void CheckForWhitespaceErrors(List<ParseError> errors, string coreContent, string extraAtEnd)
+		{
+			if (extraAtEnd.Length != 0) { errors.Add(ParseError.IllegalWhitespaceAtEnd(extraAtEnd)); }
+			if (coreContent.Contains('\t')) { errors.Add(ParseError.IllegalTabInLine()); }
+		}
+
+		private static void _ExtractCommmentRefs(List<ParseError> errors, string remainingCommentToParse, List<int> comments)
+		{
+			var formatErrorFound = false;
+			var references = remainingCommentToParse.Split(
+				new[]
+				{
+					", "
+				},
+				StringSplitOptions.None);
+			foreach (var match in references.Select(reference => Pattern.Match(reference)))
+			{
+				if (!match.Success)
+				{
+					formatErrorFound = true;
+					break;
+				}
+				comments.Add(int.Parse(match.Groups[1].Value));
+			}
+			if (formatErrorFound) { errors.Add(ParseError.IncorrectCommentFormat(remainingCommentToParse)); }
+		}
+
+		[NotNull] private static readonly Regex Pattern = new Regex(@"^\[([0-9]+)\]$");
+	}
+}
